Just to be sure, cast to Int64, hopefully fixes problem with > 4 GB files
svn path=/trunk/extragear/network/ktorrent/; revision=643778

+++ libktorrent/torrent/cachefile.cpp
@@ -335,7 +335,7 @@ namespace bt
 		}
 		
 		// jump to right position
-		SeekFile(fd,off,SEEK_SET);
+		SeekFile(fd,(Int64)off,SEEK_SET);
 		if ((Uint32)::read(fd,buf,size) != size)
 		{
 			if (close_again)
@@ -374,7 +374,7 @@ namespace bt
 		}
 		
 		// jump to right position
-		SeekFile(fd,off,SEEK_SET);
+		SeekFile(fd,(Int64)off,SEEK_SET);
 		int ret = ::write(fd,buf,size);
 		if (close_again)
 			closeTemporary();
